+++ lib/hue.rb
@@ -2,6 +2,7 @@ require 'hue/version'
 require 'hue/errors'
 require 'hue/client'
 require 'hue/bridge'
+require 'hue/editable_state'
 require 'hue/translate_keys'
 require 'hue/light'
 require 'hue/group'
